requirements.txt
# sqlite3  # built-in
pyyaml
networkx
numpy
sentence-transformers  # optional, for embeddings
requests  # optional, for ollama
pytest
